+++ lib/kennel.rb
@@ -10,6 +10,7 @@ require "kennel/progress"
 require "kennel/filter"
 require "kennel/parts_serializer"
 require "kennel/projects_provider"
+require "kennel/attribute_differ"
 require "kennel/syncer"
 require "kennel/id_map"
 require "kennel/api"
